[cli] add support back for owner field when cred checking

+++ packages/expo-cli/src/credentials/api.ts
@@ -6,6 +6,7 @@ import assign from 'lodash/assign';
 import get from 'lodash/get';
 import pick from 'lodash/pick';
 
+import invariant from 'invariant';
 import log from '../log';
 import * as appleApi from '../appleApi';
 import {
@@ -14,15 +15,25 @@ import {
   IosDistCredentials,
   IosPushCredentials,
 } from './credentials';
+import { Context } from './context';
 
 export class IosApi {
   api: ApiV2;
+  username: string;
   credentials: IosCredentials;
   shouldRefetch: boolean = true;
 
   constructor(user: User) {
     this.api = ApiV2.clientForUser(user);
+    this.username = user.username;
     this.credentials = { appCredentials: [], userCredentials: [] };
+  }
+
+  static withProjectContext(ctx: Context): IosApi {
+    invariant(ctx.hasProjectContext, 'Project context required');
+    const api = new IosApi(ctx.user);
+    api.username = ctx.manifest.owner ?? api.username;
+    return api;
   }
 
   async getAllCredentials(): Promise<IosCredentials> {
@@ -34,7 +45,7 @@ export class IosApi {
 
   async _fetchAllCredentials() {
     log('Fetching available credentials');
-    this.credentials = await this.api.getAsync('credentials/ios');
+    this.credentials = await this.api.getAsync('credentials/ios', { owner: this.username });
     this.shouldRefetch = false;
   }
 
@@ -61,7 +72,10 @@ export class IosApi {
   }
 
   async createDistCert(credentials: appleApi.DistCert): Promise<IosDistCredentials> {
-    const { id } = await this.api.postAsync('credentials/ios/dist', { credentials });
+    const { id } = await this.api.postAsync('credentials/ios/dist', {
+      credentials,
+      owner: this.username,
+    });
     const newDistCert: IosDistCredentials = { ...credentials, id, type: 'dist-cert' };
     this.credentials.userCredentials.push(newDistCert);
     return newDistCert;
@@ -72,12 +86,15 @@ export class IosApi {
   ): Promise<IosDistCredentials> {
     const { id } = await this.api.putAsync(`credentials/ios/dist/${credentialsId}`, {
       credentials,
+      owner: this.username,
     });
     const updatedDistCert: IosDistCredentials = { ...credentials, id, type: 'dist-cert' };
     const credIndex = findIndex(this.credentials.userCredentials, ({ id }) => id === credentialsId);
     this.credentials.userCredentials[credIndex] = updatedDistCert;
     return updatedDistCert;
   }
+
+  // TODO
   async deleteDistCert(credentialsId: number) {
     await this.api.deleteAsync(`credentials/ios/dist/${credentialsId}`);
     this.credentials.userCredentials = this.credentials.userCredentials.filter(
@@ -95,6 +112,7 @@ export class IosApi {
       experienceName,
       bundleIdentifier,
       userCredentialsId,
+      username: this.username,
     });
     this._ensureAppCredentials(experienceName, bundleIdentifier);
     const credIndex = findIndex(
@@ -105,7 +123,10 @@ export class IosApi {
   }
 
   async createPushKey(credentials: appleApi.PushKey): Promise<IosPushCredentials> {
-    const { id } = await this.api.postAsync('credentials/ios/push', { credentials });
+    const { id } = await this.api.postAsync('credentials/ios/push', {
+      credentials,
+      username: this.username,
+    });
     const newPushKey: IosPushCredentials = { ...credentials, id, type: 'push-key' };
     this.credentials.userCredentials.push(newPushKey);
     return newPushKey;
@@ -123,7 +144,7 @@ export class IosApi {
     return updatedPushKey;
   }
   async deletePushKey(credentialsId: number) {
-    await this.api.deleteAsync(`credentials/ios/push/${credentialsId}`);
+    await this.api.deleteAsync(`credentials/ios/push/${credentialsId}`, { owner: this.username });
     this.credentials.userCredentials = this.credentials.userCredentials.filter(
       ({ id }) => id !== credentialsId
     );
@@ -160,6 +181,7 @@ export class IosApi {
       experienceName,
       bundleIdentifier,
       userCredentialsId,
+      owner: this.username,
     });
     this._ensureAppCredentials(experienceName, bundleIdentifier);
     const credIndex = findIndex(
@@ -185,6 +207,7 @@ export class IosApi {
     await this.api.postAsync(`credentials/ios/pushCert/delete`, {
       experienceName,
       bundleIdentifier,
+      owner: this.username,
     });
     const credIndex = findIndex(
       this.credentials.appCredentials,
@@ -206,6 +229,7 @@ export class IosApi {
       experienceName,
       bundleIdentifier,
       credentials: { ...provisioningProfile, teamId: appleTeam.id },
+      owner: this.username,
     });
     const credIndex = findIndex(
       this.credentials.appCredentials,
@@ -248,6 +272,7 @@ export class IosApi {
     await this.api.postAsync(`credentials/ios/provisioningProfile/delete`, {
       experienceName,
       bundleIdentifier,
+      owner: this.username,
     });
     const credIndex = findIndex(
       this.credentials.appCredentials,
